+++ crates/plugin/src/lib.rs
@@ -1,5 +1,4 @@
 use std::io::{self, Read};
-use std::process;
 
 use javy_plugin_api::javy::Runtime;
 use javy_plugin_api::{import_namespace, Config};
@@ -8,34 +7,7 @@ use crate::shared_config::SharedConfig;
 
 mod shared_config;
 
-wit_bindgen::generate!({ world: "javy-default-plugin", generate_all });
-
 import_namespace!("javy-default-plugin-v2");
-
-struct Component;
-
-impl Guest for Component {
-    fn config_schema() -> Vec<u8> {
-        shared_config::config_schema()
-    }
-
-    fn invoke(bytecode: Vec<u8>, function: Option<String>) {
-        javy_plugin_api::invoke(&bytecode, function.as_deref()).unwrap_or_else(|e| {
-            eprintln!("{e}");
-            process::abort();
-        })
-    }
-
-    fn compile_src(src: Vec<u8>) -> Result<Vec<u8>, String> {
-        javy_plugin_api::compile_src(&src).map_err(|e| e.to_string())
-    }
-
-    fn initialize_runtime() {
-        javy_plugin_api::initialize_runtime(config, modify_runtime).unwrap()
-    }
-}
-
-export!(Component);
 
 fn config() -> Config {
     // Read shared config JSON in from stdin.
@@ -44,10 +16,9 @@ fn config() -> Config {
     // variable in subsequent invocations so a different value can't be used to
     // initialize a runtime with a different configuration.
     let mut config = Config::default();
-    config
-        .text_encoding(true)
-        .javy_stream_io(true)
-        .simd_json_builtins(true);
+    config.text_encoding(true).simd_json_builtins(true);
+    #[cfg(all(target_family = "wasm", target_os = "wasi", target_env = "p1"))]
+    config.javy_stream_io(true);
 
     let mut config_bytes = vec![];
     let shared_config = match io::stdin().read_to_end(&mut config_bytes) {
@@ -63,4 +34,9 @@ fn config() -> Config {
 
 fn modify_runtime(runtime: Runtime) -> Runtime {
     runtime
+}
+
+#[export_name = "initialize-runtime"]
+fn initialize_runtime() {
+    javy_plugin_api::initialize_runtime(config, modify_runtime).unwrap();
 }